Revert "improve StreamsBar performace by separating component & change update…"
