+++ src/pages/Timeline.tsx
@@ -1,16 +1,21 @@
-import React from 'react';
-import { List, Divider, Box } from '@mui/material';
+import React, { useContext, useState } from 'react';
+import { lighten, Paper, List, Divider, Box, Stack, TextField, Button, useTheme, IconButton, InputBase } from '@mui/material';
 import { Tweet } from '../components/Tweet'
 import { RTMMessage, StreamElement, User } from '../model';
 import { IuseResourceManager } from '../hooks/useResourceManager';
 import { IuseObjectList } from '../hooks/useObjectList';
+import { Sign } from '../util'
+import { ApplicationContext } from '../App';
+import ExploreIcon from '@mui/icons-material/Explore';
+import SearchIcon from '@mui/icons-material/Search';
+import SendIcon from '@mui/icons-material/Send';
 import { Draft } from '../components/Draft';
-import { StreamsBar } from '../components/StreamsBar';
 
 export interface TimelineProps {
     messages: IuseObjectList<StreamElement>;
     userDict: IuseResourceManager<User>;
     messageDict: IuseResourceManager<RTMMessage>;
+    inspect: (message: RTMMessage | null) => void;
     currentStreams: string;
     setCurrentStreams: (streams: string) => void;
     reload: () => void;
@@ -18,12 +23,47 @@ export interface TimelineProps {
 
 export function Timeline(props: TimelineProps) {
 
+    const appData = useContext(ApplicationContext)
+    const [draft, setDraft] = useState<string>("");
+    const theme = useTheme();
+
     return (<>
-        <StreamsBar 
-            currentStreams={props.currentStreams}
-            setCurrentStreams={props.setCurrentStreams}
-            reload={props.reload}
-        />
+        <Box sx={{
+            display: "flex",
+            justifyContent: "space-between",
+            alignItems: "center",
+            background:
+            theme.palette.primary.main,
+            padding: "5px",
+            borderRadius: "20px 20px 0 0"
+        }}>
+            <Paper
+                component="form"
+                elevation={0}
+                sx={{
+                    m: '3px 30px',
+                    p: '2px 4px',
+                    display: 'flex',
+                    alignItems: 'center',
+                    width: '100%',
+                    height: '32px',
+                    borderRadius: '16px',
+                    background: lighten(theme.palette.primary.main, 0.3)
+                }}
+            >
+                <IconButton sx={{ p: '10px' }}>
+                    <ExploreIcon sx={{ color: "white" }} />
+                </IconButton>
+                <InputBase
+                    sx={{ ml: 1, flex: 1, color: '#fff' }}
+                    value={props.currentStreams}
+                    onChange={(e) => props.setCurrentStreams(e.target.value)}
+                />
+                <IconButton sx={{ p: '10px' }} onClick={_ => props.reload()}>
+                    <SearchIcon sx={{ color: "white" }} />
+                </IconButton>
+            </Paper>
+        </Box>
         <Box sx={{overflowY: "auto", padding: "20px"}}>
             <Box>
                 <Draft currentStreams={props.currentStreams} reload={props.reload} />
@@ -36,6 +76,7 @@ export function Timeline(props: TimelineProps) {
                             message={e.Values.id}
                             messageDict={props.messageDict}
                             userDict={props.userDict}
+                            inspect={props.inspect}
                         />
                         <Divider variant="inset" component="li" sx={{margin: '0 5px'}} />
                     </React.Fragment>
